requirements.txt
asgiref==3.5.2
azure-common==1.1.27
azure-core==1.20.1
azure-nspkg==3.0.2
azure-storage-blob==12.9.0b1
backports.zoneinfo;python_version<"3.9"
boto3==1.20.4
botocore==1.23.4
cassandra-driver==3.25.0
google-cloud-core==2.2.1
google-cloud-storage==1.42.3
Django == 4.1.2
Pymongo ==3.12.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0
dnspython==2.2.1
from-root==1.0.2
numpy==1.23.4
pandas==1.5.1
plotly==5.11.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.27
mysql-connector-python==8.0.27
sqlparse==0.4.3
tenacity==8.1.0
tzdata==2022.6
openpyxl
dataflows
scikit-learn
imblearn
missingno
django-widget-tweaks
nltk
whitenoise
